Extracted logic for mapping tags and commits
