+++ core/src/main/kotlin/com/github/cdcalc/data/CommitTag.kt
@@ -1,0 +1,8 @@
+package com.github.cdcalc.data
+
+import org.eclipse.jgit.lib.ObjectId
+
+data class CommitTag (
+        val objectId: ObjectId,
+        val tagName: String
+)
