+++ core/src/test/kotlin/com/github/cdcalc/CutReleaseBranchTest.kt
@@ -1,0 +1,5 @@
+package com.github.cdcalc
+
+import org.junit.Assert.*
+
+class CutReleaseBranchTest
